Debug things

+++ src/lib.cpp
@@ -7,7 +7,7 @@ mt19937_64 e_uniform_ann;
 mt19937_64 e_uniform_pert;
 mt19937_64 e_uniform_vector;
 uniform_real_distribution<double> d_real_uniform;
-uniform_int_distribution<unsigned long long> d_int_uniform(0, 5436);
+uniform_int_distribution<int> d_int_uniform(0, 5436);
 pid_t child_pid;
 int fd[4];
